Add files via upload
This lab practices logic design using familiar building blocks. From your previous coursework, you should already be familiar with simple counters, multiplexers, decoders, and seven-segment displays. The goal of this lab is for you to implement a circuit that drives the time-multiplexed quad seven-segment display present on the Real Digital Blackboard. When you successfully complete this lab, you will have developed a piece of intellectual property that you might be able to re-use in the future.